image uplodaer fix^ question crud updated

+++ symfony/src/Service/QuestionService.php
@@ -38,17 +38,6 @@ class QuestionService
         $this->em->flush();
     }
 
-    public function imageAttach(Question $question, UploadedFile $image = null): Question
-    {
-        if ($image) {
-            $question->setImage($this->questionImageUploader->uploadImage($image, $question->getImage()));
-            $this->em->persist($question);
-            $this->em->flush();
-        };
-        return $question;
-
-    }
-
     public function saveResponse(Question $question, UploadedFile|bool|null $image): array
     {
         try {
@@ -76,29 +65,6 @@ class QuestionService
 
     }
 
-    public function saveWithVariant(Question $question, ?array $variantData, UploadedFile $questionImage = null, array $variantImages = []): Question
-    {
-        $this->imageUpdate($question, $this->questionImageUploader, $this->em, $questionImage);
-
-        $this->em->persist($question);
-
-        foreach ($variantData as $key => $variantItem) {
-            $variant = $this->variantFactory->createBuilder()->buildVariant($variantItem ?? [], $question);
-            if ($variant->getImage()) {
-                $image = array_key_exists($variant->getImage(), $variantImages) ? $variantImages[$variant->getImage()] : false;
-
-                $this->variantService->imageUpdate($variant, $this->variantImageUploader, $this->em, $image);
-            }
-
-            $this->em->persist($variant);
-
-        }
-        $this->em->flush();
-
-
-        return $question;
-    }
-
     public function saveWithVariantIfValid(Question $question, array $data, UploadedFile|bool|null $questionImage = null, array $variantImages = null, array $subtitleImages = null): array
     {
         if (!$question->getId()) {
@@ -106,10 +72,61 @@ class QuestionService
         } else {
             $message = 'Вопрос обновлен';
         }
+//        todo убрать костыль
+        if (!$question->getId()) {
+            foreach ($data['variant'] as $key => $variantData) {
+                if (key_exists($key, $variantImages)) {
+                    $data['variant'][$key]['image'] = $key;
 
+                }
+            };
+        }
+        $response = $this->createOrUpdateQuestionIfValid($question, $data, $questionImage, $variantImages, $subtitleImages);
+
+        if (!is_null($response['error'])) {
+            return [
+                'message' => 'Ошибка при вводе данных',
+                'error' => $response['error']
+            ];
+
+        } else {
+            try {
+                $question = $this->saveWithVariants($question, $response['variants'], $response['subtitles'], $questionImage, $variantImages, $subtitleImages);
+                $response = [
+                    'message' => $message,
+                    'question' => $question,
+                ];
+            } catch (\Exception $e) {
+                $response = ['error' => $e->getMessage()];
+            } catch (FilesystemException $e) {
+                $response = ['error' => $e->getMessage()];
+            } finally {
+                return $response;
+            }
+        }
+    }
+
+    public function saveWithVariants(Question $question, array $variants = [], array $subtitles = [], UploadedFile|bool|null $questionImage = null, array $variantImages = null, array $subtitleImages = null): Question
+    {
+        $this->em->persist($question);
+        $this->em->flush();
+
+        $this->imageUpdate($question, $this->questionImageUploader, $this->em, $questionImage);
+        foreach ($variants as $key => $variant) {
+            $this->variantService->imageUpdate($variant, $this->variantImageUploader, $this->em, array_key_exists($key, $variantImages) ? $variantImages[$key] : false);
+        }
+        foreach ($subtitles as $key => $subtitle) {
+            $this->variantService->imageUpdate($subtitle, $this->variantImageUploader, $this->em, array_key_exists($key, $subtitleImages) ? $subtitleImages[$key] : false);
+        }
+        return $question;
+
+    }
+
+    public function createOrUpdateQuestionIfValid(Question $question, array $data, UploadedFile|bool|null $questionImage = null, array $variantImages = null, array $subtitleImages = null): array
+    {
+        $response = [];
         $questionErrors = $this->validation->entityWithImageValidate($question, $questionImage instanceof UploadedFile ? $questionImage : null);
         $errors = $questionErrors;
-        $this->em->persist($question);
 
         $variants = [];
         $hasCorrectVariant = [];
@@ -120,9 +137,13 @@ class QuestionService
             }
             $variant = $this->variantFactory->createBuilder()->buildVariant($variantData, $question, $this->em->find(Variant::class, $key));
             if ($variant->getCorrect()) {
-                $hasCorrectVariant[] = $variant->getId();
+                $hasCorrectVariant[] = $key;
             }
-            if (!is_null($variantImages) && isset($variantImages[$key])) {
+            if (!is_null($variantImages) && isset($variantImages[$variant->getImage()])) {
+                $image = $variantImages[$variant->getImage()];
+            } else if (!is_null($variantImages) && isset($variantImages[$variant->getId()])) {
+                $image = $variantImages[$variant->getId()];
+            } else if (!is_null($variantImages) && isset($variantImages[$key])) {
                 $image = $variantImages[$key];
             } else {
                 $image = null;
@@ -132,15 +153,19 @@ class QuestionService
                 $errors[] = implode(',', $this->validation->entityWithImageValidate($variant, $image));
             } else {
                 $variants[$key] = $variant;
-                $this->em->persist($variant);
                 $question->addVariant($variant);
             }
         }
-        foreach ($question->getVariant() as $variant) {
-            if (!key_exists($variant->getId(), $variants)) {
+
+        foreach ($question->getVariant() ?? [] as $variant) {
+
+            if (!is_null($variant->getId()) && !key_exists($variant->getId(), $variants)) {
+                $question->removeVariant($variant);
                 $this->em->remove($variant);
+
             };
         }
+
         $subtitles = [];
         if (!is_null($question->getType()) && $question->getType()->getTitle() === 'conformity') {
             foreach ($data['subTitle'] ?? [] as $key => $subtitleData) {
@@ -156,9 +181,7 @@ class QuestionService
 
                 } else {
                     $subtitles[$key] = $subtitle;
-                    $this->em->persist($subtitle);
                     $question->addSubtitle($subtitle);
-
                 }
             }
         }
@@ -167,47 +190,29 @@ class QuestionService
                 $this->em->remove($subtitle);
             };
         }
-        if (count($hasCorrectVariant) < 1 && !in_array($question->getType()->getTitle(), ['input_one', 'conformity'])) {
+        if (count($hasCorrectVariant) < 1 && !in_array($question->getType()?->getTitle(), ['input_one', 'conformity'])) {
+
             $errors[] = 'Необходимо выбрать хотя бы один верный ответ';
         }
-        if (count($hasCorrectVariant) > 1 && $question->getType()->getTitle() === 'radio') {
+
+        if (count($hasCorrectVariant) > 1 && $question->getType()?->getTitle() === 'radio') {
             $errors[] = 'Необходимо выбрать только один верный ответ';
         }
 
-        if(!is_null($this->validation->uniqueTitlesValidate($question))){
-            $errors[] = implode(', ',$this->validation->uniqueTitlesValidate($question));
+        if (!is_null($this->validation->uniqueTitlesValidate($question))) {
+            $errors[] = implode(', ', $this->validation->uniqueTitlesValidate($question));
 
         }
+        if ($question->getVariant()->count() < 1) {
+            $errors[] = 'Нет соответствующих вариантов ответа (Возможные причины: пустая строка между вопросом и вариантом)';
 
-        if (!is_null($errors)) {
-            return [
-                'message' => 'Ошибка при вводе данных',
-                'error' => $errors
-            ];
-
-        } else {
-
-            try {
-                $this->em->flush();
-                $this->imageUpdate($question, $this->questionImageUploader, $this->em, $questionImage);
-                foreach ($variants as $key => $variant) {
-                    $this->variantService->imageUpdate($variant, $this->variantImageUploader, $this->em, array_key_exists($key, $variantImages) ? $variantImages[$key] : false);
-                }
-                foreach ($subtitles as $key => $subtitle) {
-                    $this->variantService->imageUpdate($subtitle, $this->variantImageUploader, $this->em, array_key_exists($key, $subtitleImages) ? $subtitleImages[$key] : false);
-                }
-                $response = [
-                    'message' => $message,
-                    'question' => $question,
-                ];
-            } catch (\Exception $e) {
-                $response = ['error' => $e->getMessage()];
-            } catch (FilesystemException $e) {
-                $response = ['error' => $e->getMessage()];
-            } finally {
-                return $response;
-            }
         }
+        $response['error'] = $errors;
+        $response['question'] = $question;
+        $response['variants'] = $variants;
+        $response['subtitles'] = $subtitles;
+
+        return $response;
     }
 
     public function getUploadedQuestionsSummary(array $questions): array
